tv_data_fetch/session_box/api_tvData/serializers.py
from rest_framework import serializers
from .models import PriceRecord

class PriceRecordSerializer(serializers.ModelSerializer):
    class Meta:
        model = PriceRecord
        fields = '__all__'
