tv_data_fetch/pyproject.toml
[project]
name = "tv-data-fetch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.16.1",
    "yfinance>=1.2.0",
]
